requirements.txt
discord.py-self>=2.0.0
